rss_loader/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
feedparser==6.0.10
frozenlist==1.3.3
idna==3.4
joblib==1.2.0
multidict==6.0.4
nltk==3.8.1
openai==0.27.8
python-dateutil==2.8.2
regex==2023.6.3
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
tiktoken==0.4.0
tqdm==4.65.0
urllib3==2.0.3
yarl==1.9.2
